+++ fiddles/angular2-cli/fiddle-0005-SlickGridTree/src/app/components/tree-grid/interfaces/index.ts
@@ -1,0 +1,4 @@
+export * from './node';
+export * from './options';
+export * from './options-change';
+