+++ fiddles/angular2-cli/fiddle-0005-SlickGridTree/src/app/components/tree-grid/interfaces/node.ts
@@ -1,0 +1,8 @@
+export interface Node {
+  id: string;
+  index: number;
+  parent?: number;
+  children?: Node[];
+  expanded?: boolean;
+  selected?: boolean;
+}