+++ fiddles/angular2-cli/fiddle-0005-SlickGridTree/src/app/components/tree-grid/interfaces/options-change.ts
@@ -1,0 +1,5 @@
+export interface OptionsChange {
+  nodes?: boolean;
+  menuOptions?: boolean;
+  filter?: boolean;
+}