+++ fiddles/angular2-cli/fiddle-0005-SlickGridTree/src/app/components/tree-grid/interfaces/options.ts
@@ -1,0 +1,8 @@
+import { Node } from './node';
+import { NavBarCmds } from '../nav-bar/enums/index';
+
+export interface Options {
+  nodes: Node[];
+  menuOptions?: NavBarCmds[];
+  filter?: string;
+}