+++ fiddles/angular2-cli/fiddle-0005-SlickGridTree/src/app/components/tree-grid/nav-bar/enums/index.ts
@@ -1,0 +1,2 @@
+export * from './nav-bar-events.enum';
+export * from './nav-bar-cmds.enum';