+++ fiddles/angular2-cli/fiddle-0005-SlickGridTree/src/app/components/tree-grid/nav-bar/enums/nav-bar-cmds.enum.ts
@@ -1,0 +1,8 @@
+export enum NavBarCmds {
+  ShowSelected,
+  ClearSelected,
+  Reload,
+  ExpandAll,
+  CollapseAll,
+  SelectAll
+}