+++ fiddles/angular2-cli/fiddle-0005-SlickGridTree/src/app/components/tree-grid/nav-bar/enums/nav-bar-events.enum.ts
@@ -1,0 +1,4 @@
+export enum NavBarEvents {
+  filter,
+  menu
+}