+++ fiddles/angular2-cli/fiddle-0005-SlickGridTree/src/app/components/tree-grid/nav-bar/index.ts
@@ -1,0 +1,4 @@
+export * from './enums/index';
+export * from './interfaces/index';
+
+