+++ fiddles/angular2-cli/fiddle-0005-SlickGridTree/src/app/components/tree-grid/nav-bar/interfaces/index.ts
@@ -1,0 +1,4 @@
+export * from './nav-bar-options';
+export * from './nav-bar-cmd';
+export * from './nav-bar-changes';
+export * from './nav-bar-event';