+++ fiddles/angular2-cli/fiddle-0005-SlickGridTree/src/app/components/tree-grid/nav-bar/interfaces/nav-bar-changes.ts
@@ -1,0 +1,5 @@
+export interface NavBarChanges {
+  isDisabled?: boolean;
+  filter?: boolean;
+  commands?: boolean;
+}