+++ fiddles/angular2-cli/fiddle-0005-SlickGridTree/src/app/components/tree-grid/nav-bar/interfaces/nav-bar-cmd.ts
@@ -1,0 +1,7 @@
+import { NavBarEvents, NavBarCmds } from '../enums/index';
+
+export interface NavBarCmd {
+  text: string;
+  event: NavBarEvents,
+  cmd: NavBarCmds
+}