+++ fiddles/angular2-cli/fiddle-0005-SlickGridTree/src/app/components/tree-grid/nav-bar/interfaces/nav-bar-event.ts
@@ -1,0 +1,6 @@
+import { NavBarEvents, NavBarCmds } from '../enums/index';
+export interface NavBarEvent {
+  event?: NavBarEvents;
+  cmd?: NavBarCmds;
+  data?: any;
+}