+++ fiddles/angular2-cli/fiddle-0005-SlickGridTree/src/app/components/tree-grid/nav-bar/interfaces/nav-bar-options.ts
@@ -1,0 +1,7 @@
+import { NavBarCmds } from '../enums/nav-bar-cmds.enum';
+
+export interface NavBarOptions {
+  isDisabled?: boolean;
+  filter?: string;
+  commands?: NavBarCmds[];
+}