+++ fiddles/angular2-cli/fiddle-0005-SlickGridTree/src/app/components/tree-grid/services/index.ts
@@ -1,0 +1,3 @@
+export * from './events.service';
+export * from './resize.service';
+export * from './node.service';