+++ fiddles/angular2-cli/fiddle-0005-SlickGridTree/src/app/components/tree-grid/tree-grid.module.ts
@@ -1,0 +1,22 @@
+import {NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
+import {ReactiveFormsModule} from '@angular/forms';
+import {TreeGridComponent} from './tree-grid.component';
+import { NavBarComponent } from './nav-bar/nav-bar.component';
+
+@NgModule({
+  imports: [
+    ReactiveFormsModule,
+    CommonModule
+  ],
+  declarations: [
+    TreeGridComponent,
+    NavBarComponent
+  ],
+  exports: [
+    TreeGridComponent
+  ]
+})
+export class TreeGridModule {
+
+}